+++ src/components/App/index.js
@@ -8,8 +8,8 @@ import Results from '../../containers/Results';
 import Experiment from '../../containers/Experiment';
 import Dashboard from '../../containers/Dashboard';
 import Downloads from '../../containers/Downloads';
+import DataSet from '../../containers/Downloads/DataSet';
 import Layout from '../Layout';
-import DataSet from '../../containers/DataSet';
 import NoMatch from '../../containers/NoMatch';
 import Privacy from '../../components/Terms/Privacy';
 import Terms from '../../components/Terms/Terms';
