+++ src/containers/Downloads/DataSet/DataSet.scss
@@ -1,4 +1,4 @@
-@import '../../common/styles/variables.scss';
+@import '../../../common/styles/variables.scss';
 
 .dataset {
   &__message {
